refactor(alerts): use context to pass dispatch
