+++ ui/src/alerting/components/notifications/StatusChangeDropdown.tsx
@@ -1,42 +1,23 @@
 // Libraries
-import React, {FC, Dispatch} from 'react'
-import {omit} from 'lodash'
+import React, {FC, useContext} from 'react'
 
 // Types
 import {StatusRuleItem} from 'src/types'
-import {Actions} from 'src/alerting/components/notifications/NewRuleOverlay.reducer'
 
 // Components
 import {Dropdown} from '@influxdata/clockface'
+import {NewRuleDispatch} from 'src/alerting/components/notifications/NewRuleOverlay'
+
+// Utils
+import {changes, changeStatusRule, activeChange} from './statusChange'
 
 interface Props {
   status: StatusRuleItem
-  dispatch: Dispatch<Actions>
 }
 
-const previousLevel = {level: 'OK'}
-const changes = ['changes from', 'equals']
+const StatusChangeDropdown: FC<Props> = ({status}) => {
+  const dispatch = useContext(NewRuleDispatch)
 
-const activeChange = status => {
-  if (!!status.value.previousLevel) {
-    return 'changes from'
-  }
-
-  return 'equals'
-}
-
-const changeStatusRule = (status, change) => {
-  if (change === 'equals' && status.value.previousLevel) {
-    return omit(status, 'value.previousLevel')
-  }
-
-  const {value} = status
-  const newValue = {...value, previousLevel}
-
-  return {...status, value: newValue}
-}
-
-const StatusChangeDropdown: FC<Props> = ({status, dispatch}) => {
   const statusChange = (s, c) =>
     dispatch({
       type: 'UPDATE_STATUS_RULES',
