+++ ui/src/alerting/components/notifications/StatusRule.tsx
@@ -1,38 +1,38 @@
 // Libraries
-import React, {FC, Dispatch} from 'react'
+import React, {FC, useContext} from 'react'
 
 // Components
 import PeriodCount from 'src/alerting/components/notifications/PeriodCount'
 import StatusLevels from 'src/alerting/components/notifications/StatusLevels'
+import {NewRuleDispatch} from 'src/alerting/components/notifications/NewRuleOverlay'
 
 // Types
 import {StatusRuleItem} from 'src/types'
-import {Actions} from 'src/alerting/components/notifications/NewRuleOverlay.reducer'
 
 interface Props {
   status: StatusRuleItem
-  dispatch: Dispatch<Actions>
 }
 
-const StatusRuleComponent: FC<Props> = ({status, dispatch}) => {
+const StatusRuleComponent: FC<Props> = ({status}) => {
+  const dispatch = useContext(NewRuleDispatch)
   const {period, count} = status.value
+
   const onChange = ({target}) => {
     const {name, value} = target
-    const newStatus = {
-      ...status,
-      [name]: value,
-    }
 
     dispatch({
       type: 'UPDATE_STATUS_RULES',
-      status: newStatus,
+      status: {
+        ...status,
+        [name]: value,
+      },
     })
   }
 
   return (
     <div className="condition-row">
       <PeriodCount period={period} count={count} onChange={onChange} />
-      <StatusLevels status={status} dispatch={dispatch} />
+      <StatusLevels status={status} />
     </div>
   )
 }
